+++ __init__.py
@@ -2,14 +2,14 @@ from typing_extensions import override
 
 from comfy_api.latest import ComfyExtension, io
 
-from .src.nodes import ZImageNegPipPrompt
+from .src.nodes import NegPipPrompt
 
 
-class NegPipZImageExtension(ComfyExtension):
+class NegPipPromptExtension(ComfyExtension):
     @override
     async def get_node_list(self) -> list[type[io.ComfyNode]]:
-        return [ZImageNegPipPrompt]
+        return [NegPipPrompt]
 
 
 async def comfy_entrypoint():
-    return NegPipZImageExtension()
+    return NegPipPromptExtension()